War Card Game Completed
Built out the logic of the game using a Game class to track the status
of a game. Every action, result of a round and the end victor is logged.
This Game object utilizes all previous objects that were built in
"War Card Game" commits so far including a Deck object that gets
shuffled, split, and dealt to two Hand objects which get assigned to
two Player objects. They all flow through the logic of the Game object.